requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
psycopg2-binary==2.9.6
Werkzeug==2.3.4
Flask-Mail
PyJWT
redis